% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_etsi_standards.R
\name{download_etsi_standards}
\alias{download_etsi_standards}
\title{Download ETSI standard documents}
\usage{
download_etsi_standards(df, path)
}
\arguments{
\item{df}{A data.frame containing column with download link \emph{PDF.link} and the name of the standard \emph{ETSI.deliverable}. This data format can be downloaded from \url{https://www.etsi.org/standards}.}

\item{path}{A string containing the path where to store the standard documents. Default is the current working directory.}
}
\description{
Download ETSI standard documents from the ETSI standards database (\url{https://www.etsi.org/standards}). as PDFs
}
