% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etsi_standards_meta.R
\docType{data}
\name{etsi_standards_meta}
\alias{etsi_standards_meta}
\title{Meta information on ETSI standards}
\format{
A data frame with 10 rows and 9 variables:
\describe{
  \item{id}{Document ID}
  \item{ETSI.deliverable}{Name of the document}
  \item{title}{Title of the document}
  \item{Status}{Status of the document}
  \item{Details.link}{Details on the standard}
  \item{PDF.link}{Download link}
  \item{Scope}{Scope / brief description}
  \item{Technical.body}{Topic of the standard}
  \item{Keywords}{Keywords describing the content of the document}
}
}
\source{
\url{https://www.etsi.org/standards/}
}
\usage{
etsi_standards_meta
}
\description{
A dataset containing meta information and download links of 10 ETSI standard
documents.
}
\keyword{datasets}
