% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_standard_doc.R
\name{parse_standard_doc}
\alias{parse_standard_doc}
\title{Standard document parser}
\usage{
parse_standard_doc(file, path, sso = "", doc.type = "pdf", 
                   overwrite = FALSE, encoding = "UTF-8", print = TRUE)
}
\arguments{
\item{file}{A character string containing the file name of the standard document.}

\item{path}{A character string containing the path of the parsed standard document.}

\item{sso}{A character string containing the acronym of a standard-setting organization (\emph{IEEE}, \emph{ETSI} or \emph{ITU-T}).}

\item{doc.type}{A character string containing the document type. Should be \emph{pdf}.}

\item{overwrite}{A logical indicating whether to overwrite existing data.}

\item{encoding}{Encoding of the input document. Default is \emph{UTF-8}. The encoding is converted to \emph{UTF-8}.}

\item{print}{A logical. If \code{TRUE} messages are printed.}
}
\description{
Parse a standard document and get the full text for each section. The parsed table of contents is stored as a csv file.
}
\seealso{
\code{\link{parse_standard_docs}}
}
