% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_standard_docs.R
\name{parse_standard_docs}
\alias{parse_standard_docs}
\title{Parse standard documents}
\usage{
parse_standard_docs(input.path, output.path, sso = "", doc.type = "pdf", 
                    overwrite = FALSE, encoding = "UTF-8", print = TRUE)
}
\arguments{
\item{input.path}{A string containing the path of the standard documents to be parsed.}

\item{output.path}{A string containing the path of the parsed standard documents.}

\item{sso}{A string containing the acronym of a standard-setting organization (\emph{IEEE}, \emph{ETSI} or \emph{ITU-T}).}

\item{doc.type}{A string containing the document type. Should be \emph{pdf}.}

\item{overwrite}{A logical indicating whether to overwrite existing data.}

\item{encoding}{Encoding of the input document. Default is \emph{UTF-8}. The encoding is converted to \emph{UTF-8}.}

\item{print}{A logical. If \code{TRUE} messages are printed.}
}
\description{
Parse standard documents and get the full text for each section. The parsed tables of contents are stored as csv's.
}
\seealso{
\code{\link{parse_standard_doc}}
}
