% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_doc_parser.R
\name{standard_doc_parser}
\alias{standard_doc_parser}
\title{Standard document parser}
\usage{
standard_doc_parser(file, path, sso = NULL, doc.type = "pdf", print = TRUE)
}
\arguments{
\item{file}{A string containing the file name of the standard document.}

\item{path}{A string containing the path of the standard document.}

\item{sso}{A string containing the acronym of a standard-setting organization (\emph{IEEE}, \emph{ETSI} or \emph{ITU-T}).}

\item{doc.type}{A string containing the document type. Should be \emph{pdf}.}

\item{print}{A logical. If \code{TRUE} messages are printed.}
}
\description{
Parses a standard document and gets the full text for each section. The parsed table of contents is stored as csv.
}
