% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_doc_parser_all.R
\name{standard_doc_parser_all}
\alias{standard_doc_parser_all}
\title{Parse all standard documents}
\usage{
standard_doc_parser_all(path, sso = NULL, doc.type = "pdf", print = TRUE)
}
\arguments{
\item{path}{A string containing the path of the standard document.}

\item{sso}{A string containing the acronym of a standard-setting organization (\emph{IEEE}, \emph{ETSI} or \emph{ITU-T}).}

\item{doc.type}{A string containing the document type. Should be \emph{pdf}.}

\item{print}{A logical. If \code{TRUE} messages are printed.}
}
\description{
Parses all standard documents and gets the full text for each section. The parsed tables of contents are stored as csv's.
}
