% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_standard_docs.R
\name{parse_standard_docs}
\alias{parse_standard_docs}
\title{Parse standard documents}
\usage{
parse_standard_docs(path, sso = NULL, doc.type = "pdf", 
       overwrite = FALSE, print = TRUE)
}
\arguments{
\item{path}{A string containing the path of the standard document.}

\item{sso}{A string containing the acronym of a standard-setting organization (\emph{IEEE}, \emph{ETSI} or \emph{ITU-T}).}

\item{doc.type}{A string containing the document type. Should be \emph{pdf}.}

\item{overwrite}{A logical indicating whether to overwrite existing data.}

\item{print}{A logical. If \code{TRUE} messages are printed.}
}
\description{
Parses standard documents and gets the full text for each section. The parsed tables of contents are stored as csv's.
}
